Renaming athletic_pandas to sweatpy
