+++ MANIFEST.in
@@ -1,1 +1,1 @@
-recursive-include athletic_pandas *.py
+recursive-include sweat *.py
