+++ Makefile
@@ -2,7 +2,7 @@ test:
 	pytest
 
 coverage:
-	coverage run --source=athletic_pandas -m pytest --
+	coverage run --source=sweat -m pytest --
 	coverage report
 	coverage html
