+++ sweat/algorithms/critical_power.py
