+++ sweat/exceptions.py
